requirements.txt
paho-mqtt==2.1.0
paho-mqtt==2.1.0
psutil==7.0.0
PyYAML==6.0.2
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
psutil==7.0.0
PyYAML==6.0.2
Werkzeug==3.1.3
